# apps/server/openapi/hunter-harness-v1.yaml
openapi: 3.1.0
info:
  title: Hunter Harness API
  version: 1.1.0
servers:
  - url: https://harness.example.com
security:
  - bearerAuth: []
paths:
...
  /api/v1/projects/{project_id}:
    get:
      operationId: getProject
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project metadata
        default:
          $ref: "#/components/responses/Error"
...
  /api/v1/ai-config/providers/{provider_id}/test:
    post:
      operationId: testAiProvider
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200": { description: AI provider connectivity test result }
        default: { $ref: "#/components/responses/Error" }
...
  /api/v1/ai-config/codex/test:
    post:
      operationId: testCodexConnection
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [schema_version]
              properties:
                schema_version: { const: 1 }
      responses:
        "200":
          { description: Codex account and selected-model connectivity result }
        "409": { description: ChatGPT account is not connected }
        default: { $ref: "#/components/responses/Error" }
...
  /api/v1/workflow-families/{slug}/sync:
    post:
      operationId: syncWorkflowFamilyFromSource
      description: Check the registered npm or GitHub source and stage newer required profiles as a draft.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Source checked; updated indicates whether a draft was staged
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkflowFamilySyncResult" }
        "409":
          {
            description: "Source content changed without a version bump (WORKFLOW_SOURCE_VERSION_CONFLICT), or the family changed during resolution (WORKFLOW_FAMILY_CHANGED)",
          }
        "422":
          {
            description: Family has no source or the source lacks required profile bundles,
          }
        "502": { description: Upstream npm or GitHub request failed }
        "504": { description: Upstream npm or GitHub request timed out }
        default: { $ref: "#/components/responses/Error" }
...
  /api/v1/projects/{project_id}/branches/{branch_name}/remote-sync/archive:prepare:
    post:
      operationId: prepareRemoteSyncArchive
      summary: Prepare a durable Remote Archive v2 operation for a stored bounded upload
      x-hunter-auth-source: authenticated_principal
      x-hunter-project-allowlist-source: server_authority
      x-hunter-project-key-scope: archive:write
      x-hunter-error-codes:
        "400": [VALIDATION_FAILED, REMOTE_ARCHIVE_INPUT_INVALID]
        "401": [AUTH_REQUIRED, TOKEN_INVALID, SESSION_INVALID]
        "403": [PROJECT_INFORMATION_FORBIDDEN, PROJECT_KEY_SCOPE, PROJECT_KEY_MISMATCH]
        "409": [REMOTE_ARCHIVE_IDEMPOTENCY_CONFLICT, REMOTE_ARCHIVE_CAPABILITY_UNAVAILABLE]
        "422": [REMOTE_ARCHIVE_PAYLOAD_HASH_MISMATCH]
        "503": [REMOTE_UNAVAILABLE]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: branch_name
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 160 }
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/RemoteSyncArchivePrepareHttpRequest" } } }
      responses:
        "201":
          description: New durable prepare claim
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
          content: { application/json: { schema: { $ref: "#/components/schemas/RemoteSyncArchivePrepareHttpResponse" } } }
        "200":
          description: Idempotent replay without a new capability claim
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
          content: { application/json: { schema: { $ref: "#/components/schemas/RemoteSyncArchivePrepareHttpResponse" } } }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /api/v1/projects/{project_id}/branches/{branch_name}/remote-sync/archive:commit:
    post:
      operationId: commitRemoteSyncArchive
      summary: Commit a prepared Remote Archive v2 operation
      x-hunter-auth-source: authenticated_principal
      x-hunter-project-allowlist-source: server_authority
      x-hunter-project-key-scope: archive:write
      x-hunter-error-codes:
        "400": [VALIDATION_FAILED, REMOTE_ARCHIVE_INPUT_INVALID]
        "401": [AUTH_REQUIRED, TOKEN_INVALID, SESSION_INVALID]
        "403": [PROJECT_INFORMATION_FORBIDDEN, PROJECT_KEY_SCOPE, PROJECT_KEY_MISMATCH]
        "404": [REMOTE_ARCHIVE_PREPARE_NOT_FOUND]
        "409": [REMOTE_ARCHIVE_LEASE_FENCED, REMOTE_ARCHIVE_LEASE_SCOPE_MISMATCH, REMOTE_ARCHIVE_CAPABILITY_UNAVAILABLE, REMOTE_ARCHIVE_COMMIT_AMBIGUOUS]
        "410": [REMOTE_ARCHIVE_PREPARE_EXPIRED]
        "503": [REMOTE_UNAVAILABLE]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: branch_name
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 160 }
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/RemoteSyncArchiveCommitHttpRequest" } } }
      responses:
        "200":
          description: Committed receipt, including a durable replay
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
          content: { application/json: { schema: { $ref: "#/components/schemas/RemoteSyncArchiveCommitHttpResponse" } } }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "410": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /api/v1/projects/{project_id}/branches/{branch_name}/remote-sync/archive/status:
    get:
      operationId: getRemoteSyncArchiveStatus
      summary: Resolve a scoped Remote Archive v2 operation
      x-hunter-auth-source: authenticated_principal
      x-hunter-project-allowlist-source: server_authority
      x-hunter-project-key-scope: archive:read
      x-hunter-error-codes:
        "400": [VALIDATION_FAILED, REMOTE_ARCHIVE_INPUT_INVALID]
        "401": [AUTH_REQUIRED, TOKEN_INVALID, SESSION_INVALID]
        "403": [PROJECT_INFORMATION_FORBIDDEN, PROJECT_KEY_SCOPE, PROJECT_KEY_MISMATCH]
        "409": [REMOTE_ARCHIVE_LEASE_SCOPE_MISMATCH]
        "503": [REMOTE_UNAVAILABLE]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: branch_name
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 160 }
        - name: operation_id
          in: query
          required: true
          schema: { type: string, pattern: "^remote_archive_operation:.{1,216}$" }
        - name: actor_id
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 160 }
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Prepared, committed, failed, or unknown durable status
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
          content: { application/json: { schema: { $ref: "#/components/schemas/RemoteSyncArchiveStatusHttpResponse" } } }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /api/v1/projects/{project_id}/branches/{branch_name}/remote-sync/archive/{operation_id}/receipt:
    get:
      operationId: getRemoteSyncArchiveReceipt
      summary: Resolve the durable receipt for a Remote Archive v2 operation
      x-hunter-auth-source: authenticated_principal
      x-hunter-project-allowlist-source: server_authority
      x-hunter-project-key-scope: archive:read
      x-hunter-error-codes:
        "400": [VALIDATION_FAILED, REMOTE_ARCHIVE_INPUT_INVALID]
        "401": [AUTH_REQUIRED, TOKEN_INVALID, SESSION_INVALID]
        "403": [PROJECT_INFORMATION_FORBIDDEN, PROJECT_KEY_SCOPE, PROJECT_KEY_MISMATCH]
        "409": [REMOTE_ARCHIVE_LEASE_SCOPE_MISMATCH]
        "503": [REMOTE_UNAVAILABLE]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: branch_name
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 160 }
        - name: operation_id
          in: path
          required: true
          schema: { type: string, pattern: "^remote_archive_operation:.{1,216}$" }
        - name: actor_id
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 160 }
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Durable receipt or null for a non-committed operation
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
          content: { application/json: { schema: { $ref: "#/components/schemas/RemoteSyncArchiveReceiptHttpResponse" } } }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
...
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        pattern: "^prj_[A-Za-z0-9_-]{1,124}$"
...
  schemas:
...
    ApiError:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
            - code
            - message
            - request_id
            - details
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
              format: uuid
            details:
              type: object
...
    WorkflowFamilySyncResult:
      type: object
      additionalProperties: false
      required: [updated, request_id]
      properties:
        updated: { type: boolean }
        version: { type: string, pattern: "^\\d+\\.\\d+\\.\\d+$" }
        request_id: { type: string, format: uuid }
...
    RemoteSyncArchiveSourceHttp:
      type: object
      additionalProperties: false
      required: [project_id, branch_name, actor_id]
      properties:
        project_id: { type: string, minLength: 1, maxLength: 160 }
        branch_name: { type: string, minLength: 1, maxLength: 160 }
        actor_id: { type: string, minLength: 1, maxLength: 160 }
        commit_sha: { type: string, minLength: 1, maxLength: 160 }
        client_id: { type: string, minLength: 1, maxLength: 160 }
        change_key: { type: string, minLength: 1, maxLength: 160 }
    RemoteSyncArchiveUploadRefHttp:
      type: object
      additionalProperties: false
      required: [ref_id, sha256, size_bytes]
      properties:
        ref_id: { type: string, minLength: 1, maxLength: 240 }
        sha256: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
        size_bytes: { type: integer, minimum: 1, maximum: 536870912 }
    RemoteSyncArchiveIdentitiesHttp:
      type: object
      additionalProperties: false
      required: [package_sha256, package_size_bytes, archive_schema_version, trusted_package_receipt_hash, local_archive_receipt_hash, manifest_hash, inventory_hash, core_v2_projection_hash]
      properties:
        package_sha256: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
        package_size_bytes: { type: integer, minimum: 1, maximum: 536870912 }
        archive_schema_version: { type: integer, const: 1 }
        trusted_package_receipt_hash: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
        local_archive_receipt_hash: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
        manifest_hash: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
        inventory_hash: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
        core_v2_projection_hash: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
    RemoteSyncArchiveMetadataHttp:
      type: object
      additionalProperties: false
      required: [schema_version, source, archive_id, identities, upload_ref]
      properties:
        schema_version: { type: integer, const: 2 }
        source: { $ref: "#/components/schemas/RemoteSyncArchiveSourceHttp" }
        archive_id: { type: string, minLength: 1, maxLength: 160 }
        identities: { $ref: "#/components/schemas/RemoteSyncArchiveIdentitiesHttp" }
        upload_ref: { $ref: "#/components/schemas/RemoteSyncArchiveUploadRefHttp" }
      x-hunter-cross-field-invariants:
        - upload_ref.sha256 and upload_ref.size_bytes equal identities.package_sha256 and identities.package_size_bytes
...
    RemoteSyncArchiveReceiptHttp:
      type: object
      additionalProperties: false
      required: [schema_version, receipt_id, operation_id, prepare_id, idempotency_key, payload_hash, source, archive_id, package_sha256, package_size_bytes, manifest_hash, trusted_package_receipt_hash, local_archive_receipt_hash, inventory_hash, core_v2_projection_hash, stored_at, receipt_hash]
      properties:
        schema_version: { type: integer, const: 2 }
        receipt_id: { type: string, pattern: "^remote_archive_receipt:sha256:[a-f0-9]{64}$" }
        operation_id: { type: string, pattern: "^remote_archive_operation:.{1,216}$" }
        prepare_id: { type: string, pattern: "^remote_archive_prepare:sha256:[a-f0-9]{64}$" }
        idempotency_key: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
        payload_hash: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
        source: { $ref: "#/components/schemas/RemoteSyncArchiveSourceHttp" }
        archive_id: { type: string, minLength: 1, maxLength: 160 }
        package_sha256: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
        package_size_bytes: { type: integer, minimum: 1, maximum: 536870912 }
        manifest_hash: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
        trusted_package_receipt_hash: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
        local_archive_receipt_hash: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
        inventory_hash: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
        core_v2_projection_hash: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
        stored_at: { type: string, format: date-time }
        receipt_hash: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
    RemoteSyncArchiveRecordHttp:
      type: object
      additionalProperties: false
      required: [schema_version, operation_id, prepare_id, idempotency_key, payload_hash, source, archive_id, identities, upload_ref, state, generation, lease, receipt, failure_code, created_at, updated_at, record_hash]
      properties:
        schema_version: { type: integer, const: 2 }
        operation_id: { type: string, pattern: "^remote_archive_operation:.{1,216}$" }
        prepare_id: { type: string, pattern: "^remote_archive_prepare:sha256:[a-f0-9]{64}$" }
        idempotency_key: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
        payload_hash: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
        source: { $ref: "#/components/schemas/RemoteSyncArchiveSourceHttp" }
        archive_id: { type: string, minLength: 1, maxLength: 160 }
        identities: { $ref: "#/components/schemas/RemoteSyncArchiveIdentitiesHttp" }
        upload_ref: { $ref: "#/components/schemas/RemoteSyncArchiveUploadRefHttp" }
        state: { type: string, enum: [pending, prepared, committing, committed, failed] }
        generation: { type: integer, minimum: 0 }
        lease:
          oneOf: [{ $ref: "#/components/schemas/RemoteSyncArchiveLeaseHttp" }, { type: "null" }]
        receipt:
          oneOf: [{ $ref: "#/components/schemas/RemoteSyncArchiveReceiptHttp" }, { type: "null" }]
        failure_code: { type: [string, "null"] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        record_hash: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
      x-hunter-cross-field-invariants:
        - record_hash is the canonical record hash without record_hash
        - prepared/committing/committed records have a lease; failed records clear lease and receipt
        - committed records have a receipt and generation at least 3
    RemoteSyncArchiveClaimHttp:
      type: object
      additionalProperties: false
      required: [operation_id, prepare_id, source, generation, fencing_token, capability]
      properties:
        operation_id: { type: string, pattern: "^remote_archive_operation:.{1,216}$" }
        prepare_id: { type: string, pattern: "^remote_archive_prepare:sha256:[a-f0-9]{64}$" }
        source: { $ref: "#/components/schemas/RemoteSyncArchiveSourceHttp" }
        generation: { type: integer, minimum: 1 }
        fencing_token: { type: integer, minimum: 1 }
        capability: { type: string, pattern: "^remote_archive_capability:.{1,216}$" }
    RemoteSyncArchivePrepareHttpRequest:
      type: object
      additionalProperties: false
      required: [schema_version, operation_id, idempotency_key, payload_hash, lease_ttl_ms, metadata]
      properties:
        schema_version: { type: integer, const: 2 }
        operation_id: { type: string, pattern: "^remote_archive_operation:.{1,216}$" }
        idempotency_key: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
        payload_hash: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
        lease_ttl_ms: { type: integer, minimum: 1, maximum: 600000 }
        metadata: { $ref: "#/components/schemas/RemoteSyncArchiveMetadataHttp" }
    RemoteSyncArchivePrepareHttpResponse:
      type: object
      additionalProperties: false
      required: [outcome, claim, record]
      properties:
        outcome: { type: string, enum: [new, replay] }
        claim:
          oneOf: [{ $ref: "#/components/schemas/RemoteSyncArchiveClaimHttp" }, { type: "null" }]
        record: { $ref: "#/components/schemas/RemoteSyncArchiveRecordHttp" }
      x-hunter-cross-field-invariants:
        - new has a claim; replay has null claim
    RemoteSyncArchiveCommitHttpRequest:
      type: object
      additionalProperties: false
      required: [claim, idempotency_key, payload_hash]
      properties:
        claim: { $ref: "#/components/schemas/RemoteSyncArchiveClaimHttp" }
        idempotency_key: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
        payload_hash: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
    RemoteSyncArchiveCommitHttpResponse:
      type: object
      additionalProperties: false
      required: [outcome, record, receipt]
      properties:
        outcome: { type: string, enum: [new, replay] }
        record: { $ref: "#/components/schemas/RemoteSyncArchiveRecordHttp" }
        receipt: { $ref: "#/components/schemas/RemoteSyncArchiveReceiptHttp" }
    RemoteSyncArchiveStatusHttpResponse:
      type: object
      additionalProperties: false
      required: [operation_id, state, record]
      properties:
        operation_id: { type: string, pattern: "^remote_archive_operation:.{1,216}$" }
        state: { type: string, enum: [pending, prepared, committing, committed, failed, unknown] }
        record:
          oneOf: [{ $ref: "#/components/schemas/RemoteSyncArchiveRecordHttp" }, { type: "null" }]
    RemoteSyncArchiveReceiptHttpResponse:
      type: object
      additionalProperties: false
      required: [receipt]
      properties:
        receipt:
          oneOf: [{ $ref: "#/components/schemas/RemoteSyncArchiveReceiptHttp" }, { type: "null" }]
  headers:
    RequestId:
      description: Stable request correlation identity echoed on every response.
      schema: { type: string, format: uuid }
  responses:
    Error:
      description: Standard error envelope
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"